+++ discord/commands/core.py
@@ -726,6 +726,8 @@ class SlashCommand(ApplicationCommand):
 
     type = 1
 
+    parent: SlashCommandGroup | None
+
     def __new__(cls, *args, **kwargs) -> SlashCommand:
         self = super().__new__(cls)
 
@@ -1119,7 +1121,7 @@ class SlashCommand(ApplicationCommand):
                 ctx.value = op.get("value")
                 ctx.options = values
 
-                if len(inspect.signature(option.autocomplete).parameters) == 2:
+                if option.autocomplete._is_instance_method:
                     instance = getattr(option.autocomplete, "__self__", ctx.cog)
                     result = option.autocomplete(instance, ctx)
                 else: