+++ CHANGELOG.md
@@ -30,16 +30,42 @@ These changes are available on the `master` branch, but have not yet been releas
   - Adds pre-typed and pre-constructed with select_type `ui.Select` aliases for the
     different select types: `ui.StringSelect`, `ui.UserSelect`, `ui.RoleSelect`,
     `ui.MentionableSelect`, and `ui.ChannelSelect`.
+- Added `store` parameter to `View` and `Modal` classes.
+  ([#2904](https://github.com/Pycord-Development/pycord/pull/2904/))
+- Added `Webhook.parent` and `Webhook.from_interaction`
+  ([#2904](https://github.com/Pycord-Development/pycord/pull/2904/))
+- Added the ability to use functions with any number of optional arguments and functions
+  returning an awaitable as `Option.autocomplete`.
+  ([#2914](https://github.com/Pycord-Development/pycord/pull/2914))
 - Added `ui.FileUpload` for modals and the `FileUpload` component.
   ([#2938](https://github.com/Pycord-Development/pycord/pull/2938))
 - Added `Attachment.read_chunked` and optional `chunksize` argument to `Attachment.save`
   for retrieving attachments in chunks.
   ([#2956](https://github.com/Pycord-Development/pycord/pull/2956))
+- Added `Permissions.bypass_slowmode`.
+  ([#2939](https://github.com/Pycord-Development/pycord/pull/2939))
 - Added support for Guild Incidents via `Guild.incidents_data` and
   `Guild.modify_incident_actions()`.
   ([#2955](https://github.com/Pycord-Development/pycord/pull/2955))
 
 ### Changed
+
+- Overhauled support for Components V2 and new Modal components
+  ([#2904](https://github.com/Pycord-Development/pycord/pull/2904/))
+  - Revert `discord.ui.View` and `discord.ui.Modal` to 2.6.1 behavior; not compatible
+    with new features.
+  - Implemented `discord.ui.DesignerView` and `discord.ui.DesignerModal` to support new
+    components.
+  - `DesignerView` and `Container` do not support `Button` and `Select` directly; use
+    `discord.ui.ActionRow` instead.
+  - `DesignerModal` does not support `InputText` and `Select` directly; use
+    `discord.ui.Label` instead.
+  - Removed `InputText.description`, `Select.label` and `Select.description`; these are
+    now attributes of `Label`.
+  - `discord.ui.Item` is now a base class for `ViewItem` and `ModalItem`; all items
+    inherit from these.
+  - All view and modal classes now inherit from a base `ItemInterface` class, split into
+    `BaseView` and `BaseModal`
 
 ### Fixed
 
@@ -66,6 +92,9 @@ These changes are available on the `master` branch, but have not yet been releas
   parameters. ([#2952](https://github.com/Pycord-Development/pycord/pull/2952))
 - Fixed autocomplete crashing when using an async staticmethod.
   ([#2966](https://github.com/Pycord-Development/pycord/pull/2966))
+- Fixed attributes like :attr:`Member.display_banner` being `None` when the member has
+  no guild specific banner, but does have a global one.
+  ([#2968](https://github.com/Pycord-Development/pycord/pull/2949))
 
 ### Deprecated
 
@@ -81,6 +110,8 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Added
 
+- Added `positional` argument to `commands.Flag`.
+  ([#2443](https://github.com/Pycord-Development/pycord/pull/2443))
 - Added `Guild.fetch_role` method.
   ([#2528](https://github.com/Pycord-Development/pycord/pull/2528))
 - Added the following `AppInfo` attributes: `approximate_guild_count`,